% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2E_Group.R
\name{cv_fold_l2e_group}
\alias{cv_fold_l2e_group}
\title{CV fold for L2E group-lasso regression}
\usage{
cv_fold_l2e_group(i, y, X, fold, cv.args, method = "median")
}
\arguments{
\item{i}{Fold index}

\item{y}{Response vector}

\item{X}{Design matrix}

\item{fold}{Vector indicating fold membership for each observation}

\item{cv.args}{List of arguments for L2E_group}

\item{method}{Character string ("median" or "mean") to compute the objective loss}
}
\value{
A numeric vector of loss values for each lambda in lambdaSeq.
}
\description{
\code{cv_fold_l2e_group} computes the loss for one cross-validation fold.
}
\keyword{internal}
