% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2E_Sparse.R
\name{cv_fold_l2e_ncv}
\alias{cv_fold_l2e_ncv}
\title{Internal function for k-fold cross-validation}
\usage{
cv_fold_l2e_ncv(i, y, X, fold, cv.args, method = "median")
}
\arguments{
\item{i}{Integer; the index of the current fold to be used as the validation set.}

\item{y}{Response vector for the full dataset.}

\item{X}{Design matrix for the full dataset.}

\item{fold}{A vector of integers indicating which fold each observation belongs to.}

\item{cv.args}{A list of arguments to be passed to \code{L2E_sparse_ncv}.}

\item{method}{Character; the method to compute the objective loss,
either "mean" or "median" (default). "median" is recommended for robustness.}
}
\value{
A numeric vector of the computed loss for each value in the
\code{lambdaSeq} provided in \code{cv.args}.
}
\description{
\code{cv_fold_l2e_ncv} fits the L2E sparse regression model on a training fold
and evaluates the robust loss on the hold-out fold.
}
\keyword{internal}
