% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2E_Group.R
\name{CV_L2E_group}
\alias{CV_L2E_group}
\title{Cross-validation for L2E group-lasso regression}
\usage{
CV_L2E_group(
  y,
  X,
  group,
  b,
  tau,
  lambdaSeq,
  nfolds = 5,
  seed = 1234,
  penalty = "grLasso",
  method = "median",
  max_iter = 100,
  tol = 1e-04,
  trace = TRUE
)
}
\arguments{
\item{y}{Response vector}

\item{X}{Design matrix}

\item{group}{Group indicator vector}

\item{b}{Initial vector of regression coefficients (optional)}

\item{tau}{Initial precision estimate (optional)}

\item{lambdaSeq}{A decreasing sequence of tuning parameter lambda values (optional)}

\item{nfolds}{The number of cross-validation folds (default is 5)}

\item{seed}{Seed for reproducibility (optional)}

\item{penalty}{The penalty to be applied to the model. For group selection, one of grLasso, grMCP, or grSCAD}

\item{method}{Character string ("median" or "mean") to compute the objective loss}

\item{max_iter}{Maximum number of iterations for the regression update}

\item{tol}{Relative tolerance}

\item{trace}{Logical; if TRUE, prints progress for each fold}
}
\value{
Returns a list containing the CV error (cve), standard error (cvse),
the index and value of the lambda with minimum CV error, the 1-standard-error lambda,
the sequence of lambda values, and the fold assignments.
}
\description{
\code{CV_L2E_group} performs k-fold cross-validation for robust L2E regression
with a group lasso penalty.
}
