% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Points_Selection.R
\name{hellinger_distance}
\alias{hellinger_distance}
\title{Compute Hellinger Distance between two samples}
\usage{
hellinger_distance(target, source)
}
\arguments{
\item{target}{Numeric vector from the target distribution.}

\item{source}{Numeric vector from the source distribution.}
}
\value{
A scalar (numeric) representing the Hellinger distance,
ranging from 0 to 1.
}
\description{
\code{hellinger_distance} calculates the Hellinger distance between the
kernel density estimates of the target and source vectors.
}
\keyword{internal}
