% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2E_Group.R
\name{L2E_group}
\alias{L2E_group}
\title{L2E group-lasso solution path with}
\usage{
L2E_group(
  y,
  X,
  group,
  b,
  tau,
  lambdaSeq,
  penalty = "grLasso",
  max_iter = 100,
  tol = 1e-04,
  Show.Time = TRUE
)
}
\arguments{
\item{y}{Response vector}

\item{X}{Design matrix}

\item{group}{Group indicator vector}

\item{b}{Initial vector of regression coefficients (optional)}

\item{tau}{Initial precision estimate (optional)}

\item{lambdaSeq}{A decreasing sequence of tuning parameter lambda values (optional)}

\item{penalty}{The penalty to be applied to the model. For group selection, one of grLasso, grMCP, or grSCAD}

\item{max_iter}{Maximum number of iterations}

\item{tol}{Relative tolerance}

\item{Show.Time}{Logical; if TRUE, prints the computing time}
}
\value{
Returns a list with estimated Beta (matrix), Tau (vector), run time, and lambdaSeq.
}
\description{
\code{L2E_group} computes the solution path for robust L2E regression with a group lasso penalty.
}
