% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2E_Group.R
\name{l2e_regression_group_accel}
\alias{l2e_regression_group_accel}
\title{L2E group-lasso regression with}
\usage{
l2e_regression_group_accel(
  y,
  X,
  beta,
  tau,
  lambda,
  group,
  penalty,
  max_iter = 100,
  tol = 1e-04,
  Show.Time = TRUE
)
}
\arguments{
\item{y}{Response vector}

\item{X}{Design matrix}

\item{beta}{Initial vector of regression coefficients}

\item{tau}{Initial precision estimate}

\item{lambda}{Tuning parameter}

\item{group}{Group indicator vector}

\item{penalty}{The penalty to be applied to the model. For group selection, one of grLasso, grMCP, or grSCAD}

\item{max_iter}{Maximum number of iterations}

\item{tol}{Relative tolerance}

\item{Show.Time}{Logical; if TRUE, prints the computing time}
}
\value{
Returns a list containing the updated beta, tau, and number of iterations.
}
\description{
\code{l2e_regression_group} performs robust L2E regression using a group lasso penalty.
}
\keyword{internal}
