% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Points_Selection.R
\name{source_cov}
\alias{source_cov}
\title{Generate an AR(1) Covariance Matrix}
\usage{
source_cov(p, rho)
}
\arguments{
\item{p}{Integer. The number of features (dimensions).}

\item{rho}{Numeric. The correlation coefficient, typically between 0 and 1.}
}
\value{
A p-by-p numeric matrix where element (i,j) is rho^|i-j|.
}
\description{
Internal helper to simulate a p-by-p covariance matrix with an AR(1) structure.
}
