% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_Points_Selection.R
\name{source_cre}
\alias{source_cre}
\title{Simulate Multiple Source Datasets}
\usage{
source_cre(n_s, p, p_s, snr, sigma_s, Sigma_S, K, out_prop, out_sig, beta_t)
}
\arguments{
\item{n_s}{Integer. Sample size for each source dataset.}

\item{p}{Integer. Total number of features.}

\item{p_s}{Integer. Number of features affected by the model shift (delta_s).}

\item{snr}{Numeric. Standard deviation of the Gaussian noise in the response.}

\item{sigma_s}{Numeric. Standard deviation of the model shift (beta_s - beta_t).}

\item{Sigma_S}{Matrix. The covariance matrix for the design matrix X.}

\item{K}{Integer. The number of source datasets to generate.}

\item{out_prop}{Numeric. Proportion of samples in each source to be corrupted by outliers (0 to 1).}

\item{out_sig}{Numeric. Magnitude of the outlier shift relative to the maximum Y value.}

\item{beta_t}{Numeric vector. The ground truth coefficients for the target domain.}
}
\value{
A list of length K, where each element is a list containing:
\item{x}{A design matrix of size n_s by p.}
\item{y}{A response vector of length n_s.}
}
\description{
Internal helper to create a list of source datasets with covariate shift,
model shift, and response outliers.
}
