% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2E_Group.R
\name{update_beta_group}
\alias{update_beta_group}
\title{Beta update in L2E group-lasso regression}
\usage{
update_beta_group(
  y,
  X,
  beta,
  tau,
  lambda,
  penalty,
  group,
  max_iter = 100,
  tol = 1e-04
)
}
\arguments{
\item{y}{Response vector}

\item{X}{Design matrix}

\item{beta}{Initial vector of regression coefficients}

\item{tau}{Initial precision estimate}

\item{lambda}{Tuning parameter}

\item{penalty}{The penalty to be applied to the model. For group selection, one of grLasso, grMCP, or grSCAD}

\item{group}{Group indicator vector (assigning each predictor to a group)}

\item{max_iter}{Maximum number of iterations}

\item{tol}{Relative tolerance}
}
\value{
Returns a list object containing the updated beta (vector) and the number of iterations used.
}
\description{
\code{update_beta_group} updates beta for robust L2E regression using a
group lasso penalty via the grpreg package.
}
\keyword{internal}
